backend/requirements.txt
django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
mongoengine==0.29.1
openai==2.14.0
python-dotenv==1.2.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10

